fix(tensordict): rename entry point and fix Check calling
- Rename tensordict_api.py to tensordict.py per spec
- Fix Check calling to use get_backend() pattern
- Add tensorclass validation tests
- Fix spec module layout duplicate entry
- Update test imports
